refactor(studio-ai): split message and code block renderers
